Clean up Imports & Exports settings code
